backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg[binary]
django-allauth
django-cors-headers
requests
cryptography
tzdata
sqlparse
asgiref
django-environ
